[!232][XAI][BUG] Fix Predicted Token Difference Scorer token selection
# Why is the change needed?
The `target` field generated by the `OccludedSpeechToTextDataset` class currently starts each item in the batch with an eos/bos token, so it can be used for generation. However, when this field is used to select the probability of predicted tokens in the `PredictedTokenDifferenceScorer`, this leads to the selection of the wrong probabilities, since no output probability distribution is associated with the bos token at the start of the `target` field.

# What changes does the patch introduce?
This patch changes the `OccludedSpeechToTextDataset` collater so it contains two distinct fields `prev_output_tokens` (bos/eos at the beginning) and `target` (bos/eos at the end) and selects the correct one every time these fields are used somewhere in the code. In this way, the fields of `OccludedSpeechToTextDataset` are more closely aligned with those of `SpeechToTextDatasetWithSrc`.

# How was this patch tested?
UTs and manual testing.