[!227][SFM][LR] Add piecewise warmup scheduler used by OWSM
# Why is the change needed?

When training deep models with a large number of layers the learning rate has to be severely decreased
as the network does not converge otherwise. This leads to poor performance for deep networks as
attested in OWSM (https://arxiv.org/pdf/2401.16658), where a piecewise learning scheduler was 
introduced as a fix to this issue.



# What changes does the patch introduce?

Adds the piecewise lr scheduler by OWSM.

# How was this patch tested?

Added UT