[!243][BUG][STREAMING] Cap negative delays to 0 in streamLAAL calculation
# Why is the change needed?

When a word/character is produced before the actual speech is started (e.g. in case of hallucinations or "(Applause)" and similar tags), the current stream_delay and stream_elapsed can become negative in the StreamLAAL code, producing negative delays in the final computation.

# What changes does the patch introduce?

The issue is solved with this MR by taking, at each delay/elapsed computation, the maximum between 0 and the actual computation i.e., the delay is always capped to 0 if negative.

# How was this patch tested?

Ad-hoc UT + manual run.