[!253][BUG][XAI] Refactor to avoid duplicate generic explanation task
# Why is the change needed?

<!--MR [226](https://gitlab.fbk.eu/mt/fbk-fairseq/-/merge_requests/226),-->

<!--MR [235](https://gitlab.fbk.eu/mt/fbk-fairseq/-/merge_requests/235).-->

During the code review of a1af02efe25bb3549407d133970a07c1cbecdbc9 it was decided to rename the `generic` explanation task to `all_tokens`. However, this change was not properly propagated to the subsequent commit 801514e4e976d933294dd79d991776e3436904e2. This lead to having two classes with different names but the same goal, with the relevant code spread between the two of them.

# What changes does the patch introduce?

This patch moves the methods in `GenericExplanationTask` to `AllTokensExplanationTask` and deletes the first class.

# How was this patch tested?

Manual runs by @spapi