[!238][XAI] Single min max normalizer for gender explanations
# Why is the change needed?
For analyses that focus only on the spectrogram saliency maps that explain the generation of gender terms, we do not want the previous output tokens' saliency maps to influence their normalization. Currently, it is only possible to perform min-max normalization for the spectrogram and previous output tokens saliency maps jointly.

# What changes does the patch introduce?
This patch introduces a new normalizer (`gender_single_min_max`) that applies min-max normalization separately to the audio and text saliency maps.

# How was this patch tested?
UTs.