[!255][BUG][XAI] Avoid any perturbation with 0 probability and fix SLIC probability
# Why is the change needed?

The current creation of masks in filterbank perturbators does not correctly handle the (very rare) case in which the random number is 0. In this scenario, even though the probability of masking is set to 0, the input is incorrectly masked.

In addition, the logic of the masking in the SLIC perturbator is wrong as it is masking the opposite of what it should. This means that by setting the probability to 0 the perturbator always masks the input instead of never doing it, and so on. 


# What changes does the patch introduce?

Fixes the two bugs described above.

# How was this patch tested?

added UTs