[!258][XAI] Deletion metric for gender
# Why is the change needed?
So far, the deletion metric could only be computed with coarse-grained metrics such as BLEU or WER. To evaluate contrastive explanations of the translation of gender, more precise metrics working at the term level, such as coverage and gender accuracy, are needed.

# What changes does the patch introduce?
Three new metrics for gender coverage (the percentage of gender terms annotated in MuST-SHE present in one of their two versions in the model outputs), gender accuracy (the percentage of in-coverage gender terms corresponding to the correct term as annotated in MuST-She), and gender flip rate (the percentage of in-coverage gender terms for which the gender was flipped with respect to the original model prediction) and their integration with the script to compute the deletion area-under-curve (AUC).

# How was this patch tested?
UTs