[!242][SFM] Add the option of computing both ASR and ST losses in a single step
# Why is the change needed?

At the moment, if we want to train a model to perform both ASR and ST tasks, we need to load two times the same dataset but with different tgt_texts and tgt_langs. This means that we have to keep the same dataset in memory two times, wasting time and resources.

# What changes does the patch introduce?

It introduces the option to leverage src_texts and src_langs without reloading the dataset two times, and then computing the losses for both ASR and ST.

# How was this patch tested?

UTs and a manual run.