[!259][BUG][XAI] Pass flag to adjust occlusion masks length
# Why is the change needed?
In !235 (commit 801514e4), we introduced a flag `force_masks_length` which, when active, adjusts the length of the occlusion mask for the previous output tokens to match the size of the previous output token embeddings (if the masks are longer, they are truncated, if they are shorter, they are padded with ones). This can be necessary when studying gender terms if the swapped hypothesis is longer or shorter than the original one. However, with the current version of the code, the flag is not properly passed along the call stack from the `OcclusionTransformerDecoderScriptable` to the `OcclusionDecoderEmbeddingsPerturbator`.

# What changes does the patch introduce?
Adds the flag `force_masks_length` as an argument of all the necessary functions so it can reach the perturbator, where it is used.

# How was this patch tested?
Manual tests using the script `fbk-fairseq/examples/speech_to_text/generate_occlusion_explanation.py` with the task `speech_to_text_genderxai` and the scorer `gender_term_contrastive_parity_ratio`.