[!241][XAI] Different aggregators to compute word probabilities
# Why is the change needed?
When computing the probability of a word (e.g. "stata") from the probability of subword tokens (e.g. "▁sta" and "ta"), simply multiplying the probabilities assigned by the model to the subword tokens that compose it penalyses long words and does not take into consideration whether these tokens constitute a fully formed word from the cases where they are just a prefix to be continued (e.g. "▁sta" and "ta" could also become "▁sta ta le").  

# What changes does the patch introduce?
This merge request makes it possible to test different aggregation strategies found in the literature to compute word probabilities. It allows to choose between simply multiplying subword probabilities, using a length normalization like what is typically done in beam search, taking only the probability of the first subword token, and the equation using the "bug" fix for beginning-of-word marking tokenisers described in [Pimentel and Meister 2024](https://aclanthology.org/2024.emnlp-main.1020/). The latter is the default choice.

To align the implementation of different gender scorers, this patch also removes the `KLGenderScorer`, which was not in use and differed significantly from the others, as it relies on entire probability distributions rather than word probabilities to assign a score.

# How was this patch tested?
UTs